# Conflicts:
#	Project_Blind/Assets/Scripts/Utils/Define.cs

+++ Project_Blind/Assets/Scripts/Core/Components/UnitHP.cs
@@ -22,7 +22,7 @@ namespace Blind
                 RefreshHpUI.Invoke(_health, _maxHealth);
         }
 
-        public UnitHP(int maxHealth) {
+        public UnitHP(float maxHealth) {
             _maxHealth = maxHealth;
             _health = _maxHealth;
